+++ src/linarr/C_list.cpp
@@ -107,21 +107,21 @@ err_type exe_linarr_C_list
 	timing::time_point begin, end;
 	double total_elapsed = 0.0;
 
-	const uint32_t n = uG.get_num_nodes();
+	const uint64_t n = uG.get_num_nodes();
 
 	// amount of linear arrangements
-	uint32_t n_linarrs;
+	uint64_t n_linarrs;
 	fin >> n_linarrs;
 
 	// linear arrangements
 	vector<vector<node>> inv_arrs(n_linarrs, vector<node>(uG.get_num_nodes()));
 	vector<linear_arrangement> arrangements(n_linarrs, linear_arrangement(uG.get_num_nodes()));
-	uint32_t single_upper_bound;
-	vector<uint32_t> list_upper_bounds(n_linarrs, 0);
+	uint64_t single_upper_bound;
+	vector<uint64_t> list_upper_bounds(n_linarrs, 0);
 
 	for (size_t i = 0; i < n_linarrs; ++i) {
 		// read linear arrangement
-		for (uint32_t u = 0; u < uG.get_num_nodes(); ++u) {
+		for (uint64_t u = 0; u < uG.get_num_nodes(); ++u) {
 			fin >> inv_arrs[i][u];
 			arrangements[i][ inv_arrs[i][u] ] = u;
 		}
@@ -134,9 +134,9 @@ err_type exe_linarr_C_list
 		fin >> single_upper_bound;
 	}
 
-	const vector<uint32_t> uCbfs = num_crossings_brute_force(uG, arrangements);
-	const vector<uint32_t> dCbfs = num_crossings_brute_force(dG, arrangements);
-	for (uint32_t i = 0; i < n_linarrs; ++i) {
+	const vector<uint64_t> uCbfs = num_crossings_brute_force(uG, arrangements);
+	const vector<uint64_t> dCbfs = num_crossings_brute_force(dG, arrangements);
+	for (uint64_t i = 0; i < n_linarrs; ++i) {
 		if (uCbfs[i] != dCbfs[i]) {
 			cerr << ERROR << endl;
 			cerr << "    Number of crossings do not coincide" << endl;
@@ -161,7 +161,7 @@ err_type exe_linarr_C_list
 		return algorithms_C::brute_force;
 	}(proc);
 
-	vector<uint32_t> uCs, dCs;
+	vector<uint64_t> uCs, dCs;
 
 	// compute all C
 	begin = timing::now();
@@ -180,7 +180,7 @@ err_type exe_linarr_C_list
 	end = timing::now();
 	total_elapsed += timing::elapsed_milliseconds(begin, end);
 
-	for (uint32_t i = 0; i < n_linarrs; ++i) {
+	for (uint64_t i = 0; i < n_linarrs; ++i) {
 		if (dCs[i] != uCs[i]) {
 			cerr << ERROR << endl;
 			cerr << "    Number of crossings do not coincide" << endl;
@@ -198,7 +198,7 @@ err_type exe_linarr_C_list
 	// uCs == dCs
 
 	if (upper_bound_type == 0) {
-		for (uint32_t i = 0; i < n_linarrs; ++i) {
+		for (uint64_t i = 0; i < n_linarrs; ++i) {
 			if (uCbfs[i] != uCs[i]) {
 				cerr << ERROR << endl;
 				cerr << "    Number of crossings do not coincide" << endl;
@@ -215,7 +215,7 @@ err_type exe_linarr_C_list
 		}
 	}
 	else if (upper_bound_type == 1) {
-		for (uint32_t i = 0; i < n_linarrs; ++i) {
+		for (uint64_t i = 0; i < n_linarrs; ++i) {
 			if (uCbfs[i] > single_upper_bound) {
 				if (uCs[i] <= uG.get_num_edges()*uG.get_num_edges()) {
 					cerr << ERROR << endl;
@@ -247,7 +247,7 @@ err_type exe_linarr_C_list
 		}
 	}
 	else if (upper_bound_type == 2) {
-		for (uint32_t i = 0; i < n_linarrs; ++i) {
+		for (uint64_t i = 0; i < n_linarrs; ++i) {
 			if (uCbfs[i] > list_upper_bounds[i]) {
 				if (uCs[i] <= uG.get_num_edges()*uG.get_num_edges()) {
 					cerr << ERROR << endl;
